Working Convexity Measure
Added convexity_metric.jl which calculates the convexity measure and plots data.

Uses hit_and_run.jl as the input for the sample_nonconvex_space() function, which is a basic implementation of hit and run sampling algorithm.

Further improvements are needed for the hit and run implementation, however it is operable for the current test cases.